Refactor CSDL3 and Tetrix for improved audio handling and platform compatibility
- Removed PulseAudioWrapper and associated files to streamline audio management, focusing on Swift-native solutions.
- Introduced a new CPulseAudio target for Linux audio support, ensuring compatibility without relying on C modules.
- Updated SDL3Game to utilize SDL3 for rendering across all platforms, enhancing consistency and performance.
- Refined VSCode launch configurations for better cross-platform development experience, including Windows-specific paths.
- Enhanced text rendering logic to utilize SDL3, simplifying the rendering process and improving maintainability.

+++ Sources/Tetrix/TetrisMusic.swift
@@ -22,7 +22,7 @@ struct AudioSpec {
     func toSDL() -> SDL_AudioSpec {
         var sdlSpec = SDL_AudioSpec()
         sdlSpec.freq = frequency
-        sdlSpec.format = SDL_AudioFormat(rawValue: format)
+        sdlSpec.format = SDL_AudioFormat(rawValue: Int32(format))
         sdlSpec.channels = Int32(channels)
         return sdlSpec
     }
